fix: certificate generation route handler
Certificate Generation Route Handler now 1 endpoint (supports query params for formats of pdf and png)
